moved dag conf module
