+++ dags/grobid_train_evaluate_grobid_train_tei.py
@@ -5,7 +5,7 @@ from sciencebeam_airflow.dags.dag_ids import (
     DEFAULT_EVALUATE_TASKS
 )
 
-from sciencebeam_dag_conf import (
+from sciencebeam_airflow.dags.dag_conf import (
     ScienceBeamDagConf,
     ScienceBeamTrainDagConf,
     ScienceBeamDatasetDagConf,
