+++ tests/grobid_train_evaluate_grobid_train_tei_test.py
@@ -3,7 +3,7 @@ from sciencebeam_airflow.dags.dag_ids import (
     DEFAULT_EVALUATE_TASKS
 )
 
-from dags.sciencebeam_dag_conf import (
+from sciencebeam_airflow.dags.dag_conf import (
     ScienceBeamDagConf,
     ScienceBeamTrainDagConf,
     ScienceBeamDatasetDagConf,
