+++ tests/grobid_train_evaluate_source_dataset_test.py
@@ -6,7 +6,7 @@ from sciencebeam_airflow.dags.dag_ids import (
     DEFAULT_CONVERT_AND_EVALUATE_TASKS
 )
 
-from dags.sciencebeam_dag_conf import (
+from sciencebeam_airflow.dags.dag_conf import (
     ScienceBeamDagConf,
     ScienceBeamTrainDagConf,
     ScienceBeamDatasetDagConf,
